refactor: separate reducer files
