+++ src/reducers/cursorBeforeSearch.js
@@ -1,0 +1,3 @@
+export const cursorBeforeSearch = ({ value }) => ({
+  cursorBeforeSearch: value
+})
