+++ src/reducers/dragInProgress.js
@@ -1,0 +1,3 @@
+export const dragInProgress = ({ value }) => ({
+  dragInProgress: value
+})
