+++ src/reducers/editing.js
@@ -1,0 +1,3 @@
+export const editing = ({ value }) => ({
+  editing: value
+})
