+++ src/reducers/expandContextItem.js
@@ -1,0 +1,7 @@
+import { equalItemsRanked } from '../util'
+
+export const expandContextItem = (state) => ({ itemsRanked }) => ({
+  expandedContextItem: equalItemsRanked(state.expandedContextItem, itemsRanked)
+    ? null
+    : itemsRanked
+})
