+++ src/reducers/render.js
@@ -1,0 +1,3 @@
+export const render = (state) => () => ({
+  dataNonce: state.dataNonce + 1
+})
