+++ src/reducers/search.js
@@ -1,0 +1,3 @@
+export const search = ({ value }) => ({
+  search: value
+})
