+++ src/reducers/searchLimit.js
@@ -1,0 +1,3 @@
+export const searchLimit = ({ value }) => ({
+  searchLimit: value
+})
