+++ src/reducers/selectionChange.js
@@ -1,0 +1,3 @@
+export const selectionChange = ({ focusOffset }) => ({
+  focusOffset
+})
