+++ src/reducers/showHelperIcon.js
@@ -1,0 +1,9 @@
+import { canShowHelper } from '../util'
+
+export const showHelperIcon = (state) => ({ id, data }) =>
+  canShowHelper(id, state)
+    ? {
+      showHelperIcon: id,
+      helperData: data
+    }
+  : {}
