+++ src/reducers/status.js
@@ -1,0 +1,3 @@
+export const status = ({ value }) => ({
+  status: value
+})
